ads/serializers/category.py
from rest_framework import serializers

from ads.models import Category


class CategoryCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'